backend/requirements.txt
django==5.2.6
djangorestframework==3.16.1
dj-rest-auth==7.0.1
django-cors-headers==4.8.0
psycopg2-binary==2.9.10
Pillow==11.3.0